Fix Python tools that fail to run (#1593)

+++ tools/nusense/generate_nanopb_message.py
@@ -34,7 +34,7 @@ def run(proto_file, selections, **kwargs):
     options_file_name = f"{file_name_no_ext}.options" if not proto_dir else f"{proto_dir}/{file_name_no_ext}.options"
 
     # If options were specified, write the options file to the same directory as the proto file
-    if not os.path.isfile(options_file_name):
+    if selections and not os.path.isfile(options_file_name):
         with open(options_file_name, "w") as f:
             # Parse each selection then write to the options file
             for selection in selections:
